Remove assert that any `make_temp_channel())` contains `python`
It is just completely baffling to me.

+++ conda_env/installers/pip.py
@@ -8,6 +8,10 @@ import os.path as op
 from conda._vendor.auxlib.compat import Utf8NamedTemporaryFile
 import tempfile
 from conda_env.pip_util import pip_subprocess
+from logging import getLogger
+
+
+log = getLogger(__name__)
 
 
 def _pip_install_via_requirements(prefix, specs, args, *_, **kwargs):
@@ -45,7 +49,11 @@ def _pip_install_via_requirements(prefix, specs, args, *_, **kwargs):
         # Win/Appveyor does not like it if we use context manager + delete=True.
         # So we delete the temporary file in a finally block.
         if requirements is not None and op.isfile(requirements.name):
-            os.remove(requirements.name)
+            if not 'CONDA_TEST_SAVE_TEMPS' in os.environ:
+                os.remove(requirements.name)
+            else:
+                log.info('CONDA_TEST_SAVE_TEMPS :: retaining pip requirements.txt {}'.format(requirements.name))
+
 
 
 # Conform to Installers API
